requirements.txt
discord
requests
python-dotenv
pillow
matplotlib
aioftp
asyncssh
psutil
aiofiles
inky
wakeonlan
beautifulsoup4
Flask[async]